+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/bindingadapter/RoomPagerBindingAdapter.kt
@@ -3,7 +3,7 @@ package com.digginroom.digginroom.feature.room.bindingadapter
 import androidx.databinding.BindingAdapter
 import com.digginroom.digginroom.feature.room.RoomEventListener
 import com.digginroom.digginroom.feature.room.customview.roominfoview.ShowRoomInfoListener
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentsListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.ShowCommentsListener
 import com.digginroom.digginroom.feature.room.customview.roompager.PagingOrientation
 import com.digginroom.digginroom.feature.room.customview.roompager.RoomPager
 import com.digginroom.digginroom.feature.room.customview.roomplayer.RoomState
