+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/DefaultShowRoomInfoListener.kt
@@ -1,7 +1,6 @@
 package com.digginroom.digginroom.feature.room.customview.roominfoview
 
 import com.digginroom.digginroom.feature.room.RoomActivity
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.TrackInfoDialog
 import com.digginroom.digginroom.model.TrackModel
 
 class DefaultShowRoomInfoListener(private val dialog: TrackInfoDialog, val activity: RoomActivity) :
