+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/RoomInfoView.kt
@@ -5,7 +5,7 @@ import android.view.LayoutInflater
 import androidx.constraintlayout.widget.ConstraintLayout
 import com.digginroom.digginroom.databinding.RoomInfoBinding
 import com.digginroom.digginroom.feature.room.RoomEventListener
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentsListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.ShowCommentsListener
 import com.digginroom.digginroom.feature.room.customview.roominfoview.navigator.DefaultRoomNavigator
 import com.digginroom.digginroom.model.CommentModel
 import com.digginroom.digginroom.model.RoomModel
