+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/adapter/CommentAdapter.kt
@@ -2,7 +2,7 @@ package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.a
 
 import android.view.ViewGroup
 import androidx.recyclerview.widget.ListAdapter
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentMenuListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.ShowCommentMenuListener
 import com.digginroom.digginroom.model.CommentModel
 
 class CommentAdapter : ListAdapter<CommentModel, CommentViewHolder>(CommentDiffUtilCallback()) {
