+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/dialog/CommentMenuDialog.kt
@@ -5,7 +5,7 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import com.digginroom.digginroom.databinding.DialogCommentMenuLayoutBinding
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.CommentMenuEventListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.CommentMenuEventListener
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 class CommentMenuDialog : BottomSheetDialogFragment() {
