+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/dialog/listener/CommentEventListener.kt
@@ -1,4 +1,4 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener
 
 interface CommentEventListener {
     fun findComments()
