+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/dialog/listener/ShowCommentsListener.kt
@@ -1,0 +1,6 @@
+package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener
+
+fun interface ShowCommentsListener {
+
+    fun show(roomId: Long)
+}
