--- android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roominfoview/comment/listener/ShowCommentsListener.kt
@@ -1,6 +1,0 @@
-package com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener
-
-fun interface ShowCommentsListener {
-
-    fun show(roomId: Long)
-}
