+++ android/app/src/main/java/com/digginroom/digginroom/feature/room/customview/roomplayer/YoutubeRoomPlayer.kt
@@ -10,7 +10,7 @@ import com.digginroom.digginroom.feature.room.RoomEventListener
 import com.digginroom.digginroom.feature.room.customview.RoomPlayerThumbnail
 import com.digginroom.digginroom.feature.room.customview.roominfoview.RoomInfoView
 import com.digginroom.digginroom.feature.room.customview.roominfoview.ShowRoomInfoListener
-import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentsListener
+import com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.ShowCommentsListener
 import com.digginroom.digginroom.model.CommentModel
 import com.digginroom.digginroom.model.RoomModel
