+++ android/app/src/main/res/layout/activity_room.xml
@@ -16,7 +16,7 @@
 
         <variable
             name="showCommentsListener"
-            type="com.digginroom.digginroom.feature.room.customview.roominfoview.comment.listener.ShowCommentsListener" />
+            type="com.digginroom.digginroom.feature.room.customview.roominfoview.comment.dialog.listener.ShowCommentsListener" />
     </data>
 
     <com.digginroom.digginroom.feature.room.customview.roompager.RoomPager
